added and fixed min rent and max rent

+++ backend/api/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Apartment(models.Model):
     city = models.CharField(max_length=100)
-    rent = models.DecimalField(max_digits=10, decimal_places=2)
+    rent = models.FloatField()
     bedrooms = models.IntegerField()
     description = models.TextField(blank=True, null=True)
     posted_date = models.DateTimeField(auto_now_add=True)
